requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
coverage==7.2.1
Django==4.1.7
djangorestframework==3.14.0
importlib-metadata==6.0.0
Markdown==3.4.1
model-bakery==1.10.1
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
zipp==3.15.0
